+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 2.4.0 (12-Sep-2025)
 - Updated SSID code (thanks, ZP!)
 - Added troubleshooting code for common JSON issues
+- Additional troubleshooting tweaks
 
 ### 2.3.0 (26-Aug-2025)
 - Enhanced `operationMode` to verbosely execute when set to `debug` (Addresses Issue #28)